imx8qm_arm2: Add xen boot commands
These commands try to be as generic as possible and avoid duplicating
argument values used for regular boot.

The fdt_file parameter still needs to be set to a .dtb for dom0. Booting
xen with stock dtb would require more hacks, for example lpuart
interrupt-parent = <&wu> needs to be removed.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>