MLK-18568 imx8: disable the DM devices removing
In u-boot 2018, all DM devices will be removed before jumping to kernel.
It executes after the board_quiesce_devices where we turn off the power
domain of the devices used in u-boot. Thus, this cause crash when removing
DM devies.

This patch disables the CONFIG_DM_DEVICE_REMOVE, since it is useless for imx8
and is not compatibled with current power domain implemenation.

Signed-off-by: Ye Li <ye.li@nxp.com>
Acked-by: Peng Fan <peng.fan@nxp.com>