MLK-18623-1 imx8mm_evk: Fix NOC and NOC APB clocks issue
Some debug codes in DDR init are used, which set NOC and NOC APB clocks to 25M
OSC and causes slow booting and poor performance in SPL and u-boot. Change to
original codes which only set DRAM APB clock to 200Mhz and use clock API to
implement it.

Some trailing spaces are also trimmed in this patch.

Signed-off-by: Ye Li <ye.li@nxp.com>