MLK-19049 mx6qsabreauto: Fix kernel NAND boot issue in bootargs
When enabling NAND boot, the ubi.mtd should be set to 6 because the
one mtdpart is added for tee.

Also the MFG_NAND_PARTITION is duplicated in mx6qsabreauto.h and
mx6sabre_common.h. We can remove it from mx6qsabreauto.h, since only
the sabreauto board have NAND.

Signed-off-by: Ye Li <ye.li@nxp.com>