MLK-20664-3 imx8m: Enable redundant offset for SPL NAND FIT
The new uuu will change to burn entire boot image to nandfit mtdpart not
only the FIT image. We enable REDUND offset configuration here, so that SPL will
try to find FIT header at two locations: the begin of nandfit mtdpart and
the 0x57c00 offset of nandfit mtdpart. This helps to be compatible with both
old and new uuu.

Signed-off-by: Ye Li <ye.li@nxp.com>