board: fsl: lx2160ardb: add QSFP+ detection support
Add QSFP transceiver detection at boot time. Transceiver model,
serial number and part number are displayed if module is detected.
An environment variable used by the Cortina 40G PHY is set, forcing
the initialization based on cable type.

The detection mechanism function is added into board_init(), thus
ensuring that we properly detect the medium side of the QSFP with DM_ETH
enabled or not. Also, the function call is under a LX2160ARDB ifdef
since we only need this mechanism on this specific board.

Signed-off-by: Florin Chiculita <florinlaurentiu.chiculita@nxp.com>
Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>