MA-14273-1 imx8q: Move spl to dram to fix reboot issue
The spl is located at ocram which will be lost when suspend/resume,
partition reboot will make A core reboot from the start address of
ATF, which blocks dual bootloader feature on Android Auto.

Move the spl to dram and make the partition reboot from spl to fix this
issue. The memory map will be changed to:
  SPL:    0x0000_0000 --------> 0x8000_0000
  ATF:    0x8000_0000 --------> 0x8004_0000
  U-boot: 0x8002_0000 --------> 0x8006_0000

Test: reboot ok on imx8qm/imx8qxp.

Change-Id: I49e8be51bf427dd00eb7507de22a787c6f6e1582
Signed-off-by: Ji Luo <ji.luo@nxp.com>