MLK-16049 imx8qm/qxp: Add support for reserving DDR memory for M4
We assign the DDR memory from 0x88000000 to 0x8FFFFFFF to M4 on QM and QXP.
The M4 can allocate this memory by two ways, in SCD or u-boot.

In this patch, u-boot addes the memory reserve node to DTB to pass the info to kernel,
no matter the M4 memory is reserved in SCD or u-boot. So kernel won't access M4 reserved memory.

Signed-off-by: Ye Li <ye.li@nxp.com>