MLK-21837 clk: imx8: Update imx8 clock driver
Update imx8 clock driver to support LPCG and full clocks tree for some
modules aligned with kernel. And add the imx8qm clock tree.

We divide the clock to serveral types: slice, fixed, lpcg, gpr, mux.
Generally slice and fixed clocks are the sources. lpcg, gpr and mux are
the downstream of those sources and are used for gating, muxing or dividing
functions.

Notice: since the clock depends on the power domain of its resource, must power
on the resource firstly, then we can get clock. Otherwise, we can't access lpcg.
Thus, the clock dump only dumps the slice clock.

Signed-off-by: Ye Li <ye.li@nxp.com>