MLK-21660 imx8qm: Disable SATA driver by default
The imx8 SATA driver will cause crash when some gen3 SATA
device is connected. As there is no booting from SATA supported by ROM,
so enabling this driver is not necessary, change to disable it at default.
And will fix the issue in future.

Users who need to use the driver, have to set "CONFIG_SATA_IMX=y" in defconfig
to enable it.

Signed-off-by: Ye Li <ye.li@nxp.com>