MLK-21885-2 lmb: refactor lmb to handle overlap in lmb reserve
lmb reserve is used to reserve some memory so that when loading images
(like kernel, dtb, initrd), images won't be loaded into the reserved memory.

The problem in current lmb is it does not handle the overlap case. When adding
a new reserved memory, if the memory region is overlap with regions already been
added in lmb, it will fail. One example is reserved memory in DTB may overlap with
u-boot relocate address. lmb reserves the u-boot relocate address firstly, so when
adding reserved memory from DTB, we will meet failure.

Actually if we handle the overlap case, we can resolve the overlap by using a max
common region for the overlap regions. So that this case won't fail.

However, u-boot has some places to use the lmb_reserve when allocating memory in
loading images.  If we allowed overlap in this function, it means images loading
address can overlap each other and cause the address check mechanism not work.

So in short, lmb_reserve has two use case:
1. Reserve memory to avoid being alloc from lmb memory. For this case,
   the reserved memory is allowed to overlap with current reserved memory
   and should coalesce with adjacent reserved memory.
2. Reserve memory to check if the memory is alreay reserved. If true,
   should return failure, then the memory can't not be used. This way
   is used in "fs", "load" command to avoid write to u-boot reserved memory.

From the two typical use cases, the functionality of lmb_reserve
is conflict. We add a new API lmb_reserve_nonoverlap for 2nd case and update
lmb_reserve for first case to handle overlap.

Signed-off-by: Ye Li <ye.li@nxp.com>