MLK-21942 DTS: imx8mq: Fix USB alias issue
Since we set the driver's compatible string to dwc3 nodes, should
use the dwc3 nodes as the USB alias, otherwise the alias won't work.
If first usb controller is probed failed, the second controller still
use 0 as seq and cause issue to the driver.

Because alias requires different names on the nodes, in the patch we have
to change the nodes name to dwc3_0 and dwc3_1

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>