MLK-22219 mx6: Place pre_misc_setting() into a common location
Currently the NoC settings are only done if the splash screen
support is enabled.

However, this can cause IPU issues on i.MX6QP in the kernel, when
splash screen is not enabled in U-Boot.

To avoid such problems, place the pre_misc_setting() function
into a common location that will always run on i.MX6QP/DP.

Signed-off-by: Fabio Estevam <fabio.estevam@nxp.com>