arm: dts: ls1028a: adds Ethernet switch node and its dependencies
The definition follows the DSA binding in kernel and describes the switch,
its ports and PHYs.
ENETC PF6 is the 2nd Eth controller linked to the switch on LS1028A, it is
not used in U-Boot and was disabled. Ethernet port aliases were also
added to better manage the multitude of ports available now, and to
enforce the order in which master and slave ports are probed.

Signed-off-by: Alex Marginean <alexandru.marginean@nxp.com>
Tested-by: Michael Walle <michael@walle.cc>
Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>