MA-16191 eliminate warning info about "is-logical" when flash images
"is-logical" is used to check whether the partition to be flashed is
dynamic partition, if it is, a partition resize command will be issued
by host fastboot. this can be well processed in fastbootd.

process this "is-logical" in uboot, now there is no logical partition,
for every partition found in u-boot, returns "no".

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>