MLK-23307-2 imx8mp: Add three reserved-memory dts nodes if inline ECC is enabled
If inline ECC is enabled on imx8mp, then the ECC region should be
reserved. Since the limit of the ddr address map when use inline ecc
and 6G memory, ECC region have to be divided into three parts. So these
three parts ECC region should be set as reserved-memory with no-map.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>