MA-16594-1 Set dtb load offset in bootloader
This commit sets the dtb load address in bootloader, the
dtb load address would be 'kernel_addr + FDT_OFFSET_TO_KERNEL'.

For imx8m, this commit will reserve 64MB for kernel, 2MB for fdt
and 16MB for ramdisk. For imx8q, this commit will reserve 60MB for
kernel, 1.5MB for fdt and 16MB for ramdisk.

This commit is a merge of below commits:
  f581f72be1 MA-16594-1 Set dtb load offset in bootloader
  207f813b2d MA-16668-2 Fix kernel hang after enabling debug configs
  42cdd5ef6d MA-16939-1 correct reserved space for kernel image for non-imx8 SoC

Change-Id: Ie597731992107ec2343e5ab6db0cab0e123f009c
Signed-off-by: Ji Luo <ji.luo@nxp.com>