LF-1229 net: eqos: invalidate RX DMA buffer in free rx pkt
When processing ICMP request, the codes use RX packet as a buffer to
re-write the mac and ip header, then copy RX packet to a TX packet.

This causes the RX packet which is actually the RX DMA buffer has dirty
data in cache. Currently the RX DMA buffer only invalidates after descriptor
has received the data. The dcache invalidate on A53 is flush + invalidate,
so it causes dirty data flushed back to the memory. Then we got the dirty
data in the reading.

Signed-off-by: Ye Li <ye.li@nxp.com>
Acked-by: Fugang Duan <fugang.duan@nxp.com>