MA-17046-1 Show orange warning for unlocked device
According to the google boot flow, an orange warning should
be displayed on UNLOCKED device to reminder the users of the
potential risks.

This commit will show an orange warning logo and warning text
on the screen, it shall be dismissed after 3 seconds, users
can also skip it by pressing the ON-OFF button.

Config 'CONFIG_AVB_WARNING_LOGO_COLS' and 'CONFIG_AVB_WARNING_LOGO_ROWS'
define the (x, y) position of the warning logo, its default
value is for 1080*720 resolution display and can be overridden.

Test: Orange warning logo show on all imx8m/imx8q platfroms.

Change-Id: I607edb3da039b47ddfac681f855834d8da187af8
Signed-off-by: Ji Luo <ji.luo@nxp.com>