MLK-23813 imx8: fuse: Skip ECC fuses check according to env variable
Currently it's not possible to use UUU to program ECC words. U-Boot is
asking for a second confirmation when programming ECC covered words.
Introduce a environment variable "force_prog_ecc", if the env is set
to "y" or "1", then the ecc fuse checking is skipped.

In uuu usage, just set this variable in uuu script, then there is no
second confirmation needed.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>