MLK-24062-5 DTS: imx8qm: Remove unused nodes from DTS
Current iMX8QM DTSi is a version with full nodes, but this introduces
overhead in u-boot initf_dm phase because the initf_dm will search all
subnodes of the root. This is does not like SPL which has slimmed the
DTB.

During initf_dm the dcache is not enabled, so the overhead is large
if we have many unused nodes in the DTS. This patch removed iMX8QM
nodes those we won't ever used in SPL/u-boot, like VPU/GPU, Audio,
camera, M4 i2c, flexcan, pwm, etc.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>