MA-17226 Invalidate the dcache after DMA operation
The main memory contents can spontaneously come to the cache due to
the speculative memory access by the CPU, this may cause coherency
problem if this happens during the DMA operaion is on-going.

Invalidate the dcache range after DMA opeartion but before the main
memory read to avoid coherency problem.

Test: reboot test.

Change-Id: I93824deab9285b5478669e0a311e0b338bf02f8a
Signed-off-by: Ji Luo <ji.luo@nxp.com>