s32v234: Update DFS_DVPORTn[MFI,MFN]
Follow the allowable values for the MFI and MFN bits, as defined in the
platform reference manual (Rev. 3, 10/2017), "Table 22-2. DFS (mfi, mfn)
settings".

Note: in Table 22-2, the DDR-DFS3 rows with PHI1 = 1066 MHz are
incorrect, because the DFS3 frequency is only allowed to be 350 or 200
MHz for the DDR-PLL, according to Table 22-1.

Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>