MLK-24426-3 imx8: Move build_info to board_late_init
Currently we call the build_info in checkboard which is a board_f
function. At this moment the u-boot env is not ready, so we can't
set any variable successfully in build_info. So moving the call to
board_late_init to fix the problem.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
(cherry picked from commit 43e1e40c6e9d65f861095989b098c2db73d11a67)