mpxs32v234: pci: fixed PCIe clocking
SBC-S32VEVB has external PCIe clocking.
Note that it is not an option to make this configuration
available via setting the environment variable 'hwconfig'.
The reason is that the variable 'hwconfig' could be deleted
by mistake from the persistent storage. In
that case U-Boot would hang when initializing the PCIe subsystem.

Signed-off-by: Costin Carabas <costin.carabas@nxp.com>