MLK-24474: Add command to set SECO FIPS mode
Add a user command to be able to program the FIPS mode.

This patchs adds the command to set it. Because the FIPS
configuration relies on fuse programming, this operation
must not be performed on any board so the command include
a check to validate the choice of the user.

The uboot is not allowed to perform the operation itself
because the SCFW API checks the caller has "SYSTEM" rights.
Instead we call the ATF which exposes a SIP call to set
the mode of the FIPS.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>