MA-17554 Decide if GKI is enabled at runtime
As we have to support GKI and non-GKI at the same time, it will be
a must to decide if the GKI is enabled or not at run-time.

This commit will decide the GKI is enabled if the parition "vendor_boot"
is found in the GPT.

This commit also make some cleanup to make the code more readable and
easier to maintain.

Test: boots on Android and Anroid Auto.

Signed-off-by: Ji Luo <ji.luo@nxp.com>
Change-Id: I6068bbaa60f5d76049b6ff0a892b5b8ca2c2f86b
(cherry picked from commit b39f7532ab524408550b8b2827bb42ab38159033)