MA-18192 Only bump the rollback index when avb verify ok
AVB verify should fail for GKI boot image but we should allow it
continue to boot in UNLOCKED state. In such case, we should not
update the stored rollback index.

This commit will update the rollback index only when the AVB
verify is OK to prevent rollback index check error.

Test: boots.

Signed-off-by: Ji Luo <ji.luo@nxp.com>
Change-Id: I82678d288edd4df6de40a1ca863ed36d3b3658a8