fit: remove an additional check for FIT images
This is a partial revert of the following commit from upstream.
commit 79af75f7776f ("fit: Don't allow verification of images with @ nodes")

This workaround is needed because the fit image for the MC firmware
image uses a firmware@1 node which describes the MC image, setup which
is not accepted starting with the commit referenced.

The MC firmware images will also be changed but this commit is also
needed so that we can use this u-boot with older MC firmware images
which cannot be changed.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>