qsfp: move QSFP detection mechanism in board_init
In the case we are running with DM_ETH, the board_eth_init() function is
not called, thus the QSFP detection mechanism is not called.
Move this portion of the code from board_eth_init() into a separate
function which will be called from board_init(), thus ensuring that we
properly detect the medium side of the QSFP.
Also, the function call is now also under a LX2160ARDB ifdef since we
only need this mechanism on this specific board.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>