MLK-25481 imx8mm/mn/mp: Keep watchdog disabled by default
WDOG HW can't be disabled once it is enabled. So after u-boot enables
the watchdog, kernel has to service the watchdog all the time.
However when suspend into s2idle mode, the WDOG counter is not stop,
then cause WDOG timeout and reset the system.

Change to not enable WDOG by default in u-boot to align previous release.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>