LF-3898-2 arm: dts: imx8mm: remove power domain property
The current power domain node in imx8mm.dtsi is not used in U-Boot,
and it not match with upstream Linux Kernel. U-Boot use arm smccc
to enable usb power domain, so we could remove this property.
It could be added back when we finally switch to upstream pgc node.

Reviewed-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>